extracted enum DeploymentFilterTypes
